♻️ refactor(routes): rename gpx feature folder to routes
Move src/features/gpx/ → src/features/routes/ (and the GpxTrackOverlay
component to RouteOverlay) and update all import paths across the
codebase. GPX is the file format; the feature is about routes more
broadly.
