+++ src/features/poster/domain/types.ts
@@ -4,7 +4,7 @@ import type {
   MarkerItem,
   MarkerProjectionInput,
 } from "@/features/markers/domain/types";
-import type { GpxTrack } from "@/features/gpx/domain/types";
+import type { GpxTrack } from "@/features/routes/domain/types";
 
 export interface CanvasSize {
   width: number;
