+++ src/features/poster/ui/PreviewPanel.tsx
@@ -9,7 +9,7 @@ import { usePosterContext } from "./PosterContext";
 import { useMapSync } from "@/features/map/application/useMapSync";
 import MapPreview from "@/features/map/ui/MapPreview";
 import MarkerOverlay from "@/features/markers/ui/MarkerOverlay";
-import GpxTrackOverlay from "@/features/gpx/ui/GpxTrackOverlay";
+import RouteOverlay from "@/features/routes/ui/RouteOverlay";
 import GradientFades from "./GradientFades";
 import PosterTextOverlay from "./PosterTextOverlay";
 import SettingsInfo from "./SettingsInfo";
@@ -413,7 +413,7 @@ export default function PreviewPanel() {
           {form.showMarkers ? (
             <GradientFades color={effectiveTheme.ui.bg} />
           ) : null}
-          <GpxTrackOverlay
+          <RouteOverlay
             tracks={state.gpxTracks}
             mapRef={mapRef}
             visible={form.showGpxTracks}
