+++ src/features/poster/ui/SettingsPanel.tsx
@@ -10,7 +10,7 @@ import LocationSection from "@/features/location/ui/LocationSection";
 import MapSettingsSection from "@/features/map/ui/MapSettingsSection";
 import LayersSection from "@/features/map/ui/LayersSection";
 import MarkersSection from "@/features/markers/ui/MarkersSection";
-import RoutesSection from "@/features/gpx/ui/RoutesSection";
+import RoutesSection from "@/features/routes/ui/RoutesSection";
 import TypographySection from "@/features/poster/ui/TypographySection";
 import {
   LocationIcon,
